# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Little Stars Daycare API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Program {
  id: number;
  name: string;
  ageRange: string;
  description: string;
  schedule: string;
  monthlyFee: number;
  capacity: number;
  enrolled: number;
  imageUrl?: string | null;
  features: string[];
  createdAt: string;
}

export interface CreateProgramBody {
  name: string;
  ageRange: string;
  description: string;
  schedule: string;
  monthlyFee: number;
  capacity: number;
  imageUrl?: string | null;
  features: string[];
}

export type EnrollmentStatus =
  (typeof EnrollmentStatus)[keyof typeof EnrollmentStatus];

export const EnrollmentStatus = {
  pending: "pending",
  approved: "approved",
  waitlisted: "waitlisted",
  rejected: "rejected",
} as const;

export interface Enrollment {
  id: number;
  parentName: string;
  parentEmail: string;
  parentPhone: string;
  childName: string;
  childAge: number;
  childDateOfBirth: string;
  programId: number;
  programName: string;
  startDate: string;
  specialNeeds?: string | null;
  status: EnrollmentStatus;
  createdAt: string;
}

export interface CreateEnrollmentBody {
  parentName: string;
  parentEmail: string;
  parentPhone: string;
  childName: string;
  childAge: number;
  childDateOfBirth: string;
  programId: number;
  startDate: string;
  specialNeeds?: string | null;
}

export interface StaffMember {
  id: number;
  name: string;
  role: string;
  bio: string;
  yearsExperience: number;
  education: string;
  imageUrl?: string | null;
  certifications: string[];
}

export type GalleryPhotoCategory =
  (typeof GalleryPhotoCategory)[keyof typeof GalleryPhotoCategory];

export const GalleryPhotoCategory = {
  activities: "activities",
  outdoor: "outdoor",
  learning: "learning",
  events: "events",
  meals: "meals",
} as const;

export interface GalleryPhoto {
  id: number;
  url: string;
  caption: string;
  category: GalleryPhotoCategory;
}

export interface Testimonial {
  id: number;
  parentName: string;
  childName: string;
  rating: number;
  review: string;
  programName: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export type EventType = (typeof EventType)[keyof typeof EventType];

export const EventType = {
  "open-house": "open-house",
  holiday: "holiday",
  workshop: "workshop",
  "field-trip": "field-trip",
  celebration: "celebration",
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  time: string;
  location: string;
  type: EventType;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  email: string;
  phone?: string | null;
  subject: string;
  message: string;
}

export interface DaycareStats {
  totalChildren: number;
  totalStaff: number;
  totalPrograms: number;
  yearsInOperation: number;
  availableSpots: number;
  totalEnrollments: number;
  pendingEnrollments: number;
  satisfactionRate: number;
}