# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Little Stars Daycare API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List all daycare programs
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Create a new program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramBody"
      responses:
        "201":
          description: Program created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get a program by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
...
components:
  schemas:
...
    Program:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ageRange:
          type: string
        description:
          type: string
        schedule:
          type: string
        monthlyFee:
          type: number
        capacity:
          type: integer
        enrolled:
          type: integer
        imageUrl:
          type: string
          nullable: true
        features:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - ageRange
        - description
        - schedule
        - monthlyFee
        - capacity
        - enrolled
        - features
        - createdAt
    CreateProgramBody:
      type: object
      properties:
        name:
          type: string
        ageRange:
          type: string
        description:
          type: string
        schedule:
          type: string
        monthlyFee:
          type: number
        capacity:
          type: integer
        imageUrl:
          type: string
          nullable: true
        features:
          type: array
          items:
            type: string
      required:
        - name
        - ageRange
        - description
        - schedule
        - monthlyFee
        - capacity
        - features
...