# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Little Stars Daycare API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all daycare programs
 */
export const ListProgramsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  ageRange: zod.string(),
  description: zod.string(),
  schedule: zod.string(),
  monthlyFee: zod.number(),
  capacity: zod.number(),
  enrolled: zod.number(),
  imageUrl: zod.string().nullish(),
  features: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});
export const ListProgramsResponse = zod.array(ListProgramsResponseItem);

/**
 * @summary Create a new program
 */
export const CreateProgramBody = zod.object({
  name: zod.string(),
  ageRange: zod.string(),
  description: zod.string(),
  schedule: zod.string(),
  monthlyFee: zod.number(),
  capacity: zod.number(),
  imageUrl: zod.string().nullish(),
  features: zod.array(zod.string()),
});

/**
 * @summary Get a program by ID
 */
export const GetProgramParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProgramResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  ageRange: zod.string(),
  description: zod.string(),
  schedule: zod.string(),
  monthlyFee: zod.number(),
  capacity: zod.number(),
  enrolled: zod.number(),
  imageUrl: zod.string().nullish(),
  features: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List enrollment requests
 */
export const ListEnrollmentsResponseItem = zod.object({
  id: zod.number(),
  parentName: zod.string(),
  parentEmail: zod.string(),
  parentPhone: zod.string(),
  childName: zod.string(),
  childAge: zod.number(),
  childDateOfBirth: zod.string(),
  programId: zod.number(),
  programName: zod.string(),
  startDate: zod.string(),
  specialNeeds: zod.string().nullish(),
  status: zod.enum(["pending", "approved", "waitlisted", "rejected"]),
  createdAt: zod.coerce.date(),
});
export const ListEnrollmentsResponse = zod.array(ListEnrollmentsResponseItem);

/**
 * @summary Submit enrollment request
 */
export const CreateEnrollmentBody = zod.object({
  parentName: zod.string(),
  parentEmail: zod.string(),
  parentPhone: zod.string(),
  childName: zod.string(),
  childAge: zod.number(),
  childDateOfBirth: zod.string(),
  programId: zod.number(),
  startDate: zod.string(),
  specialNeeds: zod.string().nullish(),
});

/**
 * @summary List all staff members
 */
export const ListStaffResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  role: zod.string(),
  bio: zod.string(),
  yearsExperience: zod.number(),
  education: zod.string(),
  imageUrl: zod.string().nullish(),
  certifications: zod.array(zod.string()),
});
export const ListStaffResponse = zod.array(ListStaffResponseItem);

/**
 * @summary List gallery photos
 */
export const ListGalleryResponseItem = zod.object({
  id: zod.number(),
  url: zod.string(),
  caption: zod.string(),
  category: zod.enum(["activities", "outdoor", "learning", "events", "meals"]),
});
export const ListGalleryResponse = zod.array(ListGalleryResponseItem);

/**
 * @summary List parent testimonials
 */
export const ListTestimonialsResponseItem = zod.object({
  id: zod.number(),
  parentName: zod.string(),
  childName: zod.string(),
  rating: zod.number(),
  review: zod.string(),
  programName: zod.string(),
  avatarUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListTestimonialsResponse = zod.array(ListTestimonialsResponseItem);

/**
 * @summary List upcoming events
 */
export const ListEventsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  date: zod.string(),
  time: zod.string(),
  location: zod.string(),
  type: zod.enum([
    "open-house",
    "holiday",
    "workshop",
    "field-trip",
    "celebration",
  ]),
});
export const ListEventsResponse = zod.array(ListEventsResponseItem);

/**
 * @summary Submit contact form
 */
export const SubmitContactBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  subject: zod.string(),
  message: zod.string(),
});

/**
 * @summary Get daycare stats
 */
export const GetStatsResponse = zod.object({
  totalChildren: zod.number(),
  totalStaff: zod.number(),
  totalPrograms: zod.number(),
  yearsInOperation: zod.number(),
  availableSpots: zod.number(),
  totalEnrollments: zod.number(),
  pendingEnrollments: zod.number(),
  satisfactionRate: zod.number(),
});