# lib/api-zod/src/generated/types/createProgramBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Little Stars Daycare API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProgramBody {
  name: string;
  ageRange: string;
  description: string;
  schedule: string;
  monthlyFee: number;
  capacity: number;
  imageUrl?: string | null;
  features: string[];
}